Changed the way we work with secrets.
SOOM_SEC is no longer needed. We just work with CUSTOM_SEC which from now on will be called soomlaSecret. This secret will be provided on initialization of Soomla meaning that the initialization process for Soomla modules is: 1. Soomla.initialize("your soomla secret here to encrypt the user data"); 2. Initialize the modules and plugins you want.